btc-signal-executor/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
alpaca-py==0.43.2
pydantic==2.12.5
python-dotenv==1.2.2
